refactor: 유저 정보 찾기 관련 공용 로직은 common 패키지로 이동함
