--- src/main/java/com/frolic/sns/auth/application/finder/FinderType.java
@@ -1,5 +1,0 @@
-package com.frolic.sns.auth.application.finder;
-
-public enum FinderType {
-  EMAIL, PASSWORD
-}
