+++ src/main/java/com/frolic/sns/auth/application/finder/common/AuthCodeCacheManager.java
@@ -1,4 +1,4 @@
-package com.frolic.sns.auth.application.finder;
+package com.frolic.sns.auth.application.finder.common;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.ValueOperations;
