+++ src/main/java/com/frolic/sns/auth/application/finder/common/FinderType.java
@@ -1,0 +1,5 @@
+package com.frolic.sns.auth.application.finder.common;
+
+public enum FinderType {
+  EMAIL, PASSWORD
+}
