+++ src/main/java/com/frolic/sns/auth/application/finder/common/UserInfoFindable.java
@@ -1,4 +1,7 @@
-package com.frolic.sns.auth.application.finder;
+package com.frolic.sns.auth.application.finder.common;
+
+import com.frolic.sns.auth.application.finder.common.AuthCode;
+import com.frolic.sns.auth.application.finder.common.FinderType;
 
 import java.util.UUID;
